+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * `wn.ic` module ([#40]
 * `wn.taxonomy` module ([#125])
+* `wn.similarity.res` Resnik similarity ([#122])
 
 ### Changed
 
@@ -420,4 +421,5 @@ abandoned, but this is an entirely new codebase.
 [#115]: https://github.com/goodmami/wn/issues/115
 [#116]: https://github.com/goodmami/wn/issues/116
 [#117]: https://github.com/goodmami/wn/issues/117
+[#122]: https://github.com/goodmami/wn/issues/122
 [#125]: https://github.com/goodmami/wn/issues/125